Major update: interface changes to juicer, CI
This commit adds the following features:
  - Progress tracking for juicer
  - Logging capabilities for juicer
  - Docstrings
  - TravisCI ready (all prerequisites install, no tests yet)
  - Incremental file update (see below)
  - Layout updates
  - new LICENSE

All output that was being printed to screen previously is either being
logged to the logfile or being suppressed as unneeded. The previous
output has been replaced by text informing the user of the progress of
the juicing.

Instead of waiting until the end to write out all the URLs at once
(which is a disaster if, e.g. someone shuts the system down), the
program now incrementally writes out the URLs as it finds them.
Uniqueness is enforced by the UniqueFileWriter class, which maintains a
record of what has already been written to the file.

Program is now TravisCI ready (phantomJS and pip requirements install
correctly) though there is no actual testing of the build yet, since
setup.py is currently an empty file.

General updates to the layout of the program, preparing it for docs and
testing suites.

Added MIT License.